backend/requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter (optional)

# Environment variables
python-dotenv>=1.0.0

# AI/ML (for future implementation)
# scikit-learn>=1.3.0
# pandas>=2.0.0
# numpy>=1.24.0

# Utilities
python-dateutil>=2.8.2

# Development
black>=23.0.0
flake8>=6.0.0
